requirements.txt
antiorm==1.2.1
appdirs==1.4.4
bcrypt==4.0.1
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
db==0.1.1
db-sqlite3==0.0.1
dnspython==2.3.0
email-validator==1.3.1
-e git+ssh://git@github.com/sabhapathi0306/fyndproject.git@14060a2fe4bdfaf4a3797f0b195197bf6804ed2e#egg=financeproject
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozendict==2.3.7
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multitasking==0.0.11
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
yfinance==0.2.17
